# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CARE-E hub API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/shipments": {
      "get": {
        "tags": [
          "shipments"
        ],
        "summary": "List Shipments",
        "description": "Shipments the caller's org is involved in (from, to or carrier), newest first; a\nLOGISTICS org also sees every unassigned (CREATED) shipment: the dispatch board.",
        "operationId": "list_shipments_api_v1_shipments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ShipmentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "assigned_to_me",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only the shipments assigned to the caller as driver.",
              "default": false,
              "title": "Assigned To Me"
            },
            "description": "Only the shipments assigned to the caller as driver."
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "inbound",
                    "outbound"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inbound: shipments to the caller's org (its deliveries); outbound: shipments from it.",
              "title": "Direction"
            },
            "description": "inbound: shipments to the caller's org (its deliveries); outbound: shipments from it."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ShipmentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shipments/{shipment_id}": {
      "get": {
        "tags": [
          "shipments"
        ],
        "summary": "Get Shipment",
        "description": "One shipment with its route geometry, status history and last driver location.\n403 unless the caller's org is involved (or it is unassigned and the caller is a\nlogistics org).",
        "operationId": "get_shipment_api_v1_shipments__shipment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shipment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shipments/{shipment_id}/assign": {
      "post": {
        "tags": [
          "shipments"
        ],
        "summary": "Assign Shipment",
        "description": "CREATED -> ASSIGNED with one of the caller's org's drivers and vehicles (403 for\nanother org's). A cold-chain shipment needs a cold-chain vehicle: 400 with the reason\notherwise. Stores the road route, its geometry and the ETA (OSRM, haversine fallback).\n409 unless CREATED.",
        "operationId": "assign_shipment_api_v1_shipments__shipment_id__assign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shipment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shipments/{shipment_id}/unassign": {
      "post": {
        "tags": [
          "shipments"
        ],
        "summary": "Unassign Shipment",
        "description": "ASSIGNED -> CREATED, by the carrier org (403 otherwise); 409 unless ASSIGNED.",
        "operationId": "unassign_shipment_api_v1_shipments__shipment_id__unassign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shipment Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReasonIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shipments/{shipment_id}/status": {
      "post": {
        "tags": [
          "shipments"
        ],
        "summary": "Update Shipment Status",
        "description": "The assigned driver only (403 otherwise): ASSIGNED -> PICKED_UP -> IN_TRANSIT ->\nDELIVERED, one step at a time (409 otherwise). PICKED_UP draws down the source's\non_hand and consumes its FIRM hold.",
        "operationId": "update_shipment_status_api_v1_shipments__shipment_id__status_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shipment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shipments/{shipment_id}/location": {
      "post": {
        "tags": [
          "shipments"
        ],
        "summary": "Post Shipment Location",
        "description": "A GPS ping from the assigned driver's phone (403 otherwise), while ASSIGNED,\nPICKED_UP or IN_TRANSIT (409 otherwise). Sent to the shipment's orgs as\n`shipment.location`.",
        "operationId": "post_shipment_location_api_v1_shipments__shipment_id__location_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shipment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drivers": {
      "get": {
        "tags": [
          "shipments"
        ],
        "summary": "List Drivers",
        "description": "The caller's own org's drivers.",
        "operationId": "list_drivers_api_v1_drivers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DriverOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles": {
      "get": {
        "tags": [
          "shipments"
        ],
        "summary": "List Vehicles",
        "description": "The caller's own org's vehicles.",
        "operationId": "list_vehicles_api_v1_vehicles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_VehicleOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_ShipmentOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ShipmentOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[ShipmentOut]"
      },
...
      "ReasonIn": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ReasonIn"
      },
...
      "ShipmentDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "shortage_id": {
            "type": "string",
            "format": "uuid",
            "title": "Shortage Id"
          },
          "source_request_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Request Id"
          },
          "purchase_order_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Order Id"
          },
          "from_org_id": {
            "type": "string",
            "format": "uuid",
            "title": "From Org Id"
          },
          "from_org_name": {
            "type": "string",
            "title": "From Org Name"
          },
          "to_org_id": {
            "type": "string",
            "format": "uuid",
            "title": "To Org Id"
          },
          "to_org_name": {
            "type": "string",
            "title": "To Org Name"
          },
          "carrier_org_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carrier Org Id"
          },
          "carrier_org_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carrier Org Name"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "product_code": {
            "type": "string",
            "title": "Product Code"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "qty": {
            "type": "integer",
            "title": "Qty"
          },
          "requires_cold_chain": {
            "type": "boolean",
            "title": "Requires Cold Chain"
          },
          "status": {
            "$ref": "#/components/schemas/ShipmentStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "required_by": {
            "type": "string",
            "format": "date-time",
            "title": "Required By",
            "description": "The shortage's deadline."
          },
          "driver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriverRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "vehicle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VehicleOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "planned_eta": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Eta"
          },
          "eta": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eta"
          },
          "route_distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Distance Km"
          },
          "route_provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RouteProvider"
              },
              {
                "type": "null"
              }
            ]
          },
          "pickup": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StopOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "drop": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StopOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "route_geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route Geometry",
            "description": "GeoJSON LineString ([lng, lat] pairs) of the road route; null until assigned."
          },
          "status_history": {
            "items": {
              "$ref": "#/components/schemas/StatusChangeOut"
            },
            "type": "array",
            "title": "Status History"
          },
          "last_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "inspection_note_required": {
            "type": "boolean",
            "title": "Inspection Note Required",
            "description": "True when the shipment has an open cold-chain excursion: its receipt then needs an inspection note (business-rules.md \u00a79). Always false until S15."
          },
          "receipt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReceiptOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "What the receiving org recorded (with the reconciliation once every shipment of the shortage has a receipt). Shown to the receiving org only; null for the other orgs and before a receipt."
          }
        },
        "type": "object",
        "required": [
          "id",
          "shortage_id",
          "source_request_id",
          "purchase_order_id",
          "from_org_id",
          "from_org_name",
          "to_org_id",
          "to_org_name",
          "carrier_org_id",
          "carrier_org_name",
          "product_id",
          "product_code",
          "product_name",
          "qty",
          "requires_cold_chain",
          "status",
          "priority",
          "required_by",
          "driver",
          "vehicle",
          "device_id",
          "planned_eta",
          "eta",
          "route_distance_km",
          "route_provider",
          "pickup",
          "drop",
          "created_at",
          "updated_at",
          "route_geometry",
          "status_history",
          "last_location",
          "inspection_note_required",
          "receipt"
        ],
        "title": "ShipmentDetailOut"
      },
...
      "ShipmentStatus": {
        "type": "string",
        "enum": [
          "CREATED",
          "ASSIGNED",
          "PICKED_UP",
          "IN_TRANSIT",
          "DELIVERED",
          "RECONCILED"
        ],
        "title": "ShipmentStatus"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}